Allow overriding default validator
When associating the field to a form, validator does not get called
unless onChange has been called at least once.

If a user submits a form without ever changing the field conent
then validator is never called and allows and empty field to
pass the form validation.

Expose the validator completely and allow developers to run
custom validation on form submissions.

Remove asynchronous validation as we are not using it.